+++ src/angular-public/autocomplete/autocomplete/autocomplete-trigger.directive.ts
@@ -19,7 +19,6 @@ import {
   ElementRef,
   forwardRef,
   Host,
-  HostBinding,
   HostListener,
   Inject,
   InjectionToken,
@@ -99,6 +98,16 @@ export const SBB_AUTOCOMPLETE_SCROLL_STRATEGY_FACTORY_PROVIDER = {
       multi: true,
     },
   ],
+  host: {
+    '[attr.role]': `this.autocompleteDisabled ? null : 'combobox'`,
+    '[attr.autocomplete]': 'this.autocompleteAttribute',
+    '[attr.aria-expanded]': 'this.autocompleteDisabled ? null : this.panelOpen.toString()',
+    '[attr.aria-owns]':
+      'this.autocompleteDisabled || !this.panelOpen ? null : this.autocomplete.id',
+    '[attr.aria-autocomplete]': `this.autocompleteDisabled ? null : 'list'`,
+    '[attr.aria-activedescendant]': 'this.activeOption ? this.activeOption.id : null',
+    '[class.sbb-autocomplete-expanded]': 'this.autocompleteDisabled ? null : this.panelOpen',
+  },
 })
 export class SbbAutocompleteTrigger implements ControlValueAccessor, AfterViewInit, OnDestroy {
   private _overlayRef: OverlayRef | null;
@@ -139,7 +148,11 @@ export class SbbAutocompleteTrigger implements ControlValueAccessor, AfterViewIn
 
   private _inputValue = new BehaviorSubject('');
 
-  @HostBinding('attr.role') get role() {
+  /**
+   * @Deprecated Internal detail
+   * @docs-private
+   */
+  get role() {
     return this._autocompleteDisabled ? null : 'combobox';
   }
 
@@ -191,7 +204,6 @@ export class SbbAutocompleteTrigger implements ControlValueAccessor, AfterViewIn
    */
   // tslint:disable-next-line:no-input-rename
   @Input('autocomplete')
-  @HostBinding('attr.autocomplete')
   autocompleteAttribute = 'off';
 
   /** Stream of autocomplete option selections. */
@@ -210,11 +222,19 @@ export class SbbAutocompleteTrigger implements ControlValueAccessor, AfterViewIn
     );
   });
 
-  @HostBinding('attr.aria-expanded') get ariaExpanded(): string | null {
+  /**
+   * @Deprecated Internal detail
+   * @docs-private
+   */
+  get ariaExpanded(): string | null {
     return this.autocompleteDisabled ? null : this.panelOpen.toString();
   }
 
-  @HostBinding('attr.aria-owns') get ariaOwns(): string | null {
+  /**
+   * @Deprecated Internal detail
+   * @docs-private
+   */
+  get ariaOwns(): string | null {
     return this.autocompleteDisabled || !this.panelOpen ? null : this.autocomplete.id;
   }
 
@@ -249,12 +269,18 @@ export class SbbAutocompleteTrigger implements ControlValueAccessor, AfterViewIn
     this._autocompleteDisabled = coerceBooleanProperty(value);
   }
 
-  @HostBinding('attr.aria-autocomplete')
+  /**
+   * @Deprecated Internal detail
+   * @docs-private
+   */
   get ariaAutocomplete(): string | null {
     return this._autocompleteDisabled ? null : 'list';
   }
 
-  @HostBinding('attr.aria-activedescendant')
+  /**
+   * @Deprecated Internal detail
+   * @docs-private
+   */
   get activeOptionId() {
     return this.activeOption ? this.activeOption.id : null;
   }